basic modules  ported to esp32 as single source for esp8266/esp32

+++ src_lfs/_init.lua
@@ -7,7 +7,20 @@
 --------------------------------------------------------------------
 -- junand 22.09.2016
 
--- dofile ( "startup.lua" );
+local node, _file = node, file;
+
+-- set file compatiblity table in global context
+file = {};
+file.open = io and io.open or _file.open;
+file.read = io and io.read or _file.read;
+file.write = io and io.write or _file.write;
+file.close = io and io.close or _file.close;
+file.exists = _file.exists;
+file.rename = _file.rename;
+file.remove = _file.remove;
+local file = file;
+print ( "_init: file=" .. tostring ( file ) .. " file.open=" .. tostring ( file.open ) );
+
 
 -------------------------------------------------------------------------------
 --  Settings
@@ -15,6 +28,8 @@
 local NO_BOOT_FILE = "no_boot";
 local LFS_RELOAD_FILE = "lfs_reload";
 local LFS_TS_FILE = "lfs.img.ts";
+
+local START_TELNET_ON_ERROR = false;
 
 ----------------------------------------------------------------------------------------
 -- private
@@ -29,43 +44,55 @@ local LFS_TS_FILE = "lfs.img.ts";
 -- 6, external reset
 local rawcode, bootreason, cause = node.bootreason ();
 
-local startTelnet;
+local startTelnet = false;
 
 --------------------------------------------------------------------
 -- public
+
+--------------------------------------------------------------------
+-- application global
+
+function unrequire ( module )
+
+    local m = package.loaded [module];
+    if ( m and m.subunrequire ) then m.subunrequire (); end
+
+    package.loaded [module] = nil
+    _G [module] = nil
+
+end
+
+function tohex ( byte, len )
+
+    return "0x" .. string.format( "%0" .. (len or 2) .. "X", byte );
+
+end
 
 package.loaders [3] = function ( module ) -- loader_flash
     local fn, ba = node.flashindex ( module );
     return ba and "Module not in LFS" or fn;
 end
 
+is_ESP32 = node.chipmodel ~= nil;
+
+--------------------------------------------------------------------
+
 print ( "[INIT] boot: rawcode=" .. rawcode .. " reason=" .. bootreason .. " cause=" .. tostring ( cause ) );
 
 
-if ( file.exists ( LFS_RELOAD_FILE ) ) then
-    file.remove ( LFS_RELOAD_FILE );
-    file.remove ( "_" .. LFS_TS_FILE );
-    node.restart ();
-    print ( "[INIT] restart after lfs reload" );
-    return;
-else
-    if ( bootreason == 1 or bootreason == 2 or bootreason == 3 ) then
-        if ( file.exists ( NO_BOOT_FILE ) ) then
-            print ( "[INIT] booting after error; NO STARTUP" );
-            startTelnet = true;
-        else
-            file.open ( NO_BOOT_FILE, "w" );
-            file.close ();
-        end
+if ( START_TELNET_ON_ERROR and ( bootreason == 1 or bootreason == 2 or bootreason == 3 ) ) then
+    if ( file.exists ( NO_BOOT_FILE ) ) then
+        print ( "[INIT] booting after error; NO STARTUP" );
+        startTelnet = true;
     else
-        if ( file.exists ( NO_BOOT_FILE ) ) then
-            file.remove ( NO_BOOT_FILE );
-        end
+        local f = file.open ( NO_BOOT_FILE, "w" );
+        f:close ();
     end
-
+else
+    if ( file.exists ( NO_BOOT_FILE ) ) then
+        file.remove ( NO_BOOT_FILE );
+    end
 end
-
---require ( "_lfs" );
 
 require ( "startup" ).start ( startTelnet );
